move check all into separate js controller
