+++ src/UDS.Net.Forms/Pages/Shared/_VisitStatusFilter.cshtml
@@ -23,27 +23,25 @@
                 </button>
             </h3>
             <!-- Filter section, show/hide based on section state. -->
-            <div data-controller="statusFilter" data-dropdown-target="menu" class="pt-6 right-24 w-52 absolute bg-white z-10" id="filter-section-mobile-0">
+            <div data-dropdown-target="menu" class="pt-6 right-24 w-52 absolute bg-white z-10" id="filter-section-mobile-0">
                 <div class="space-y-5 border border-gray-300 rounded pl-3 py-3">
-                    <div class="space-y-3">
-                        <div class="flex items-center" data-action="click->statusFilter#ToggleAllStatuses">
-                            <input data-statusFilter-target="toggleAll"  id="markAll" type="checkbox" class="size-4 rounded border-gray-300 text-indigo-600 focus:ring-indigo-500">
+                    <div data-controller="checkboxSelectAll" class="space-y-3">
+                        <div class="flex items-center" data-action="click->checkboxSelectAll#ToggleAllStatuses">
+                            <input data-checkboxSelectAll-target="toggleAll"  id="markAll" type="checkbox" class="size-4 rounded border-gray-300 text-indigo-600 focus:ring-indigo-500">
                             <label for="markAll" class="ml-3 min-w-0 flex-1 text-gray-900 font-bold">Select All</label>
                         </div>
                         @foreach (var status in Model.StatusFilter.PacketStatuses)
                         {
-                            <div class="flex items-center" data-action="click->statusFilter#CheckStatusFilterCount click->statusFilter#BuildStatusListString">
-                                <input data-statusFilter-target="statusCheckbox" id="status_[@status]" value="@status" type="checkbox" class="size-4 rounded border-gray-300 text-indigo-600 focus:ring-indigo-500">
+                            <div class="flex items-center">
+                                <input data-checkboxSelectAll-target="checkBox" id="status_[@status]" value="@status" type="checkbox" class="size-4 rounded border-gray-300 text-indigo-600 focus:ring-indigo-500">
                                 <label for="status_[@status]" class="ml-3 min-w-0 flex-1 text-gray-900">@status</label>
                             </div>
                         }
                     </div>
                     <div class="flex justify-center">
-                        <button data-statusFilter-target="apply" type="submit" class="flex-none rounded-md bg-indigo-600 px-3.5 py-2.5 text-sm font-semibold text-white shadow-sm hover:bg-indigo-500 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-indigo-600 disabled:bg-gray-400">Apply</button>
+                        <button type="submit" class="flex-none rounded-md bg-indigo-600 px-3.5 py-2.5 text-sm font-semibold text-white shadow-sm hover:bg-indigo-500 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-indigo-600 disabled:bg-gray-400">Apply</button>
                     </div>
                 </div>
-                <input disabled data-statusFilter-target="currentStatusString" class="hidden" asp-for="StatusFilter.StatusListString" />
-                <input data-statusFilter-target="newStatusString" class="hidden" name="statuses" type="text"/>
             </div>
         </div>
      </form>
